requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2018.8.24
chardet==3.0.4
coverage==4.5.1
dj-database-url==0.5.0
Django==2.1
django-contrib-comments==1.9.0
django-heroku==0.3.1
django-mptt==0.8.7
django-parler==1.9.2
django-parler-rest==1.4.2
django-polymorphic==2.0.3
django-rosetta==0.9.0
django-tinymce==2.7.0
django-treebeard==4.3
djangorestframework==3.8.2
easy-thumbnails==2.5
gunicorn==19.9.0
idna==2.7
kombu==3.0.37
Pillow==5.2.0
polib==1.1.0
psycopg2==2.7.5
pytz==2018.5
redis==2.10.6
requests==2.19.1
six==1.11.0
sorl-thumbnail==12.4.1
Unidecode==0.4.21
urllib3==1.23
vine==1.1.4
whitenoise==4.0
